confirm

+++ public/app.js
@@ -1,38 +1,54 @@
 import Index from "./js/index.js";
 import levels from "./js/levels.js";
 import createElement from "./js/createElement.js";
+import indexContext from "./contexts/indexContext.js";
+import confirm from "./js/confirm.js";
+
+
+
 function App() {
-    this.levelNo =  10;
-    this.index = new Index(this.levelNo,this.ganar.bind(this),this.fallar.bind(this));
+    this.levelNo =  localStorage.getItem("level") || 1;
+    this.index = new Index(this.levelNo);
     this.index.game.setGrid();
     this.index.control.setBackSize();
+    indexContext.set("ganar",this.ganar.bind(this));
+    indexContext.set("fallar",this.fallar.bind(this));
+
+   
 }
-App.prototype.ganar = function() {
-    setTimeout(() => {
+App.prototype.ganar = async function() {
+    setTimeout(async () => {
         this.index.control.play.playButton.click();
-        alert("Has ganado!!");
-        document.body.innerHTML = "";
-        if(this.levelNo < levels.length) {
-            this.levelNo++;
+        const res = await confirm("Has Ganado!",[{name:"Sigiente nivel",value:true},{name:"Volver",value:false}]);
+        console.log(res)
+        if(res) {
+            document.body.innerHTML = "";
+            if(this.levelNo < levels.length) {
+                this.levelNo++;
+            }else {
+                this.levelNo = 1;
+            }
+            localStorage.setItem("level",this.levelNo);
+            this.index = new Index(this.levelNo);
+            this.index.game.setGrid();
+            this.index.control.setBackSize();
         }else {
-            this.levelNo = 1;
+            document.body.innerHTML = "";
+            localStorage.setItem("level",this.levelNo); 
+            this.index = new Index(this.levelNo);
+            this.index.game.setGrid();
+            this.index.control.setBackSize();
         }
-        localStorage.setItem("level",this.levelNo);
-        this.index = new Index(this.levelNo,this.ganar.bind(this),this.fallar.bind(this));
-        this.index.game.setGrid();
-        this.index.control.setBackSize();
+       
     }, 10);
    
     
 }
-App.prototype.fallar = function() {
-    setTimeout(() => {
+App.prototype.fallar =  function() {
+    setTimeout(async () => {
         this.index.control.play.playButton.click();
-        alert("Has fallado!!");
-        this.index.game.setStars();
-        this.index.control.play.stop.click();
-        this.index.control.play.play = false;
-        this.index.game.avion.restart(this.levelNo-1);
+        const res = await confirm("has perdido!",[{name:"Volver",value:true}]);
+        this.index.control.play.stop.click(); 
     }, 10);
     
 }
